requirements.txt
streamlit
tensorflow
Pillow
numpy
streamlit-lottie
requests 